Building Docker image for Ledger
Initially I wanted this to be a more or less atomic commit, but after a
while the experimentation got out of hand. So fuck it, here you go, this
single commit adds something that can builds a Docker image which can be
used to build Ledger. I'll write a separate story about how complicated
all of this is, and how fucked up C++ and CMake are.