Main Makefile
I like using Make as a driver for builds since it's very convenient
to be able to just "make" after checkout to build something useful.

Create a main Makefile that will be controlling the build. For now,
it can only build a Docker image (or pull it from the registry).