Basic Gradle build
Set up a basic build of Android AAR with Gradle. This will compile some
native code (Hello world for now), package it into AAR, along with some
helper code that should be able to run the binary. (This is untested at
the moment.)

We'll work on using actual Ledger in the next commits.