Use stamp files correctly
Internal make targets should depend on the stamp files, not on the
phony targets. If they depend on the phony targets, they are always
not satisfied and the targets get rebuild. For example, this causes
the Docker image checks to be performed all the time, defeating any
benefit gained from the stamp files.