Ledger JNI library
I've been thinking that invoking a binary directly is too crude for
interacting with Ledger. I don't really want to parse its output if
there is an option for a programmatic API.

So! Instead of executing a new process every time we need to do
something, we will use the code from the library in-process.
Unfortunately, this code is all in C++ and Android apps don't talk C++,
they are all in Kotlin or Java. Therefore we need a JNI wrapper library
to forward JVM stuff into C++ land. This is not going to be cheap, but
who care as long as it works.

Building JNI library is simple: that's just a shared library with some
peculiar symbols. Building a library using "libledger" is harder. First,
Ledger is not really designed to be used as a library so we have to hook
in out build pretty deep in it and make sure we include all necessary
stuff. Another obstacle is linkage. If we link in static libraries, they
are going to export all their symbols. We don't want GMP and Boost crap
to be sticking out of our shared object. In order to avoid this we have
to use a "linker script" aka "version file" which tells the linker which
symbols we want to be exported and which ones we want to be hidden.

Finally, there are some things that you need to be aware of when writing
code that uses Ledger. Make sure to include <system.hh> file before any
other Ledger includes, since that's the header which is implicitly
included during regular Ledger compilation, and it pulls in all the
dependencies, such as Boost.