Basic Session class
Introduce a basic API to Ledger in form of a Session class. This
directly maps onto ledger::session_t, as close as the code allows.

Note that Ledger is not *really* tested as a library. It sure is
designed to be used as one, but there aren't many (if any) users
of it this way. Therefore there are some weird spots. One of them
is initialization of globals, which in Ledger CLI is normally performed
in global.cc, before any libledger functions are used. Moreover! In
order to perform some initialization, you need a session object.
Well, there isn't much we can do about it here...

Also, some utilities are provided for error reporting. Ledger has quite
a few exception types, but for the time being I don't particularely
care about them. JNI definitely does not care, it only needs to catch
any C++ exceptions and convert them into Java ones.