Use a submodule fork insted of patch set
Using patches is fine for Makefile-based builds but it's less convenient
for other build systems (e.g., bare Gradle on developer's machine). Thus
instead of applying patches on the fly, just switch the Ledger submodule
to a repository with a fork which already has some patches applied.

I start thinking that it's too early to split off "ledger" JNI library
into a separate repository from the main app. It made sense back when
I though that just building a command-line binary will be enough. In
that case no more development is necessary. However, a JNI binding is
going to be in constant development and having it separate during the
initial prototyping stage is going to be very painful. Therefore, this
repository is going to slowly mutate into a main one.