Android Studio stuff
Commit stuff from ".idea" directory to preserve project setup. Note that
I use auto-importing of Gradle project, and Gradle build files are the
source of truth here.

Unfortunately, I find Android Studio pretty much unusable of my laptop
because of the sheer amount of C++ crap in the project. It takes ages
for Android Studio to index all the Boost include shit, it just has to
reindex it every time there is a change in the build system or C++
files, and you're stuck waiting for "Updating symbols" to complete,
because Android Studio will refuse to build before that. (It's a known
"bug", people have been asking to fix it since at least 2017.)

Maybe my desktop machine will be better off with this since it has 32 GB
of RAM and a better CPU, but I have to say that for laptop development
I guess I'll stick to two-project split after all.