Support older CMakes (< 3.13)
Android Studio uses an older version of CMake from Android SDK which
does not understand target_link_options(). Replace it with a hack by
linking the binary to a 'library' which actually is linker flags.
This works for GCC/Clang as target_link_libraries() just pastes its
arguments into the command line.