Require CMake provided by Android SDK
Instead of using lax requirements on CMake version, use the same version
as provided by Android SDK. It is recent enough to suit our needs, and
we can include it into the base image. (For some reason Gradle likes to
ignore the system-installed CMake if there is no explicit "version"
clause in build.gradle.)

Note that we still need to build a newer CMake from source in order to
build fucking^W Boost. That thing does not go away.