Update session_ptr interface
Make the session_ptr actually useful. It will look like a smart pointer,
but actually it's a 'dumb pointer' -- the one which does not really
manage memory. However, you can still use "->" to call methods of
ledger::session_t. For this sake another wrapper is introduced which
actually forwards calls.