Improve incremental builds of amalgamation library
While working on the updates I have noticed that "ar" does not rebuild
the amalgamation library if it sees that it's already in place and all
the component libraries are present. This can lead to pretty surprising
behavior when you make a change but it does not get to the combined
library. Fix this up by always removing the library if we are going to
rebuild it with "ar".

Also, add a "ranlib" pass to reindex object files and symbols after
building the amalgamation. Normally GNU "ar" does it by itself, but
not in the MRI mode which we use. Indexing is important for link-time
optimizations later.