Enable link-time optimizations
We use a wrapper library to forward the calls to actual Ledger. This
adds unnecessary function calls in the generated assembly. Make sure
that link-time optimizations are enabled so that the linker can inline
those wrapper calls and avoid an unnecessary indirection.

During testing I have found that Gradle likes to pass "-Oz" flag to
linker when LTO is enabled. This works on some platforms and does not
work on the others (ARMv7, for example). Make sure that we use some
numeric optimization level supported by the linker.

Also, don't enable LTO in debug builds, for the sake of convenience
and useful source-line information.

And finally, Instead of using a weird hack to add linker flags, add them
directly to CMake property of the target. Modern CMake versions have
a dedicated function for that, but we invent our own since we need to
support and use an older version of CMake. Because Boost and Android.