Build Ledger for both release and debug mode
Since LTO actually produces LLVM IR, the JNI wrapper has be using LTO if
Ledger wrapper uses it. In case of a debugging build we'd been to build
the wrapper in debug mode as well.

Therefore, produce two builds of Ledger amalgamation. Note that this
affects only Ledger source code and the wrapper. The dependencies like
Boost and GMP are used as is (they are compiled in release mode).