Relicense under GNU LGPL v3+
It seems that I have went a bit overboard with using GNU GPL for a library.
I'd still like it to be reusable by other applications if they choose to.
Many people are afraid of GPL so let them make a choice.

As noted in the README, the resulting Ledger AAR is a combined work of the
following parts:

  - Java code providing JVM interface to Ledger facilities
  - JNI code gluing JVM world with native code
  - Ledger wrapper library insulating JNI code from Boost madness
  - Ledger CLI, embedded as a library
  - Boost libraries which are dependencies of Ledger
  - GNU MPFR library which is a dependency of Ledger
  - GNU MP library which is a dependency of Ledger and GNU MPFR
  - parts of Bionic C library included everywhere as a result of
    targeting Android NDK environment
  - libc++ provided by Android NDK is also embedded with C++ code

Java code, JNI code, and Ledger wrapper source code is available in this
repository. It is distributed under GNU LGPL v3 or later, just like the
combined Ledger AAR.

Ledger CLI code includes some modifications made by me, and is distributed
under its original BSD-style license. It is compatible with GNU LGPL v3.

Boost libraries are redistributed intact, under their original Boost
software license 1.0. It is compatible with GNU LGPL v3.

GNU MPFR is distributed under GNU LGPL v3 or later, and redistributed here
under the same terms.

GNU MP is distributed under GNU LGPL v3 or GNU GPL v2, with provisions for
redistributions under any later version of those licenses. We make use of
these provisions to redistribute it under GNU LGPL v3 or any later version.

Google's Bionic is distributed under BSD-style license. It is not directly
included into the binaries, but parts of its header files are used during
compilation, and the resulting objects are linked against it. In any way,
this combination is compatible with GNU LGPL v3.

Similarly, all C++ code in the combined binary uses and includes libc++
with the C++ Standard Library provided by Android NDK. It is redistributed
under the terms of MIT license or BSD-style UIUC license, either of which
is compatible with GNU LGPL v3.

Note that the build scripts -- Makefile, CMakeLists.txt, and assorted shell
scripts -- are all distributed under MIT (aka Expat) license as well. This
should not be that significant since they are unlikely to be a part of the
software distribution, but just in case I am using permissive license here.